% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{Any}
\alias{Any}
\title{Cloud Tool Results API Objects 
Reads and publishes results from Firebase Test Lab.}
\usage{
Any(typeUrl = NULL, value = NULL)
}
\arguments{
\item{typeUrl}{A URL/resource name whose content describes the type of the serialized protocol buffer message}

\item{value}{Must be a valid serialized protocol buffer of the above specified type}
}
\value{
Any object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:20:23
filename: /Users/mark/dev/R/autoGoogleAPI/googletoolresultsv1beta3.auto/R/toolresults_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Any Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Any` contains an arbitrary serialized protocol buffer message along with a URL that describes the type of the serialized message.Protobuf library provides support to pack/unpack Any values in the form of utility functions or additional generated methods of the Any type.Example 1: Pack and unpack a message in C++.Foo foo = ...; Any any; any.PackFrom(foo); ... if (any.UnpackTo(&foo)) { ... }Example 2: Pack and unpack a message in Java.Foo foo = ...; Any any = Any.pack(foo); ... if (any.is(Foo.class)) { foo = any.unpack(Foo.class); }Example 3: Pack and unpack a message in Python.foo = Foo(...) any = Any() any.Pack(foo) ... if any.Is(Foo.DESCRIPTOR): any.Unpack(foo) ...The pack methods provided by protobuf library will by default use 'type.googleapis.com/full.type.name' as the type URL and the unpack methods only use the fully qualified type name after the last '/' in the type URL, for example 'foo.bar.com/x/y.z' will yield type name 'y.z'.JSON ==== The JSON representation of an `Any` value uses the regular representation of the deserialized, embedded message, with an additional field `@type` which contains the type URL. Example:package google.profile; message Person { string first_name = 1; string last_name = 2; }{ '@type': 'type.googleapis.com/google.profile.Person', 'firstName': , 'lastName':  }If the embedded message type is well-known and has a custom JSON representation, that representation will be embedded adding a field `value` which holds the custom JSON in addition to the `@type` field. Example (for message [google.protobuf.Duration][]):{ '@type': 'type.googleapis.com/google.protobuf.Duration', 'value': '1.212s' }
}
