% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{Duration}
\alias{Duration}
\title{Duration Object}
\usage{
Duration(nanos = NULL, seconds = NULL)
}
\arguments{
\item{nanos}{Signed fractions of a second at nanosecond resolution of the span of time}

\item{seconds}{Signed seconds of the span of time}
}
\value{
Duration object
}
\description{
Duration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Duration represents a signed, fixed-length span of time represented as a count of seconds and fractions of seconds at nanosecond resolution. It is independent of any calendar and concepts like 'day' or 'month'. It is related to Timestamp in that the difference between two Timestamp values is a Duration and it can be added or subtracted from a Timestamp. Range is approximately +-10,000 years.Example 1: Compute Duration from two Timestamps in pseudo code.Timestamp start = ...; Timestamp end = ...; Duration duration = ...;duration.seconds = end.seconds - start.seconds; duration.nanos = end.nanos - start.nanos;if (duration.seconds  0) { duration.seconds += 1; duration.nanos -= 1000000000; } else if (durations.seconds > 0 && duration.nanos < 0) { duration.seconds -= 1; duration.nanos += 1000000000; }Example 2: Compute Timestamp from Timestamp + Duration in pseudo code.Timestamp start = ...; Duration duration = ...; Timestamp end = ...;end.seconds = start.seconds + duration.seconds; end.nanos = start.nanos + duration.nanos;if (end.nanos = 1000000000) { end.seconds += 1; end.nanos -= 1000000000; }Example 3: Compute Duration from datetime.timedelta in Python.td = datetime.timedelta(days=3, minutes=10) duration = Duration() duration.FromTimedelta(td)
}
\seealso{
Other res: \code{\link{Any}}, \code{\link{Execution}},
  \code{\link{FailureDetail}}, \code{\link{FileReference}},
  \code{\link{History}}, \code{\link{Image}},
  \code{\link{InconclusiveDetail}},
  \code{\link{ListExecutionsResponse}},
  \code{\link{ListHistoriesResponse}},
  \code{\link{ListStepThumbnailsResponse}},
  \code{\link{ListStepsResponse}}, \code{\link{Outcome}},
  \code{\link{ProjectSettings}},
  \code{\link{PublishXunitXmlFilesRequest}},
  \code{\link{SkippedDetail}}, \code{\link{StackTrace}},
  \code{\link{Status}},
  \code{\link{StepDimensionValueEntry}},
  \code{\link{StepLabelsEntry}}, \code{\link{Step}},
  \code{\link{SuccessDetail}},
  \code{\link{TestCaseReference}},
  \code{\link{TestExecutionStep}}, \code{\link{TestIssue}},
  \code{\link{TestSuiteOverview}},
  \code{\link{TestTiming}}, \code{\link{Thumbnail}},
  \code{\link{Timestamp}}, \code{\link{ToolExecutionStep}},
  \code{\link{ToolExecution}}, \code{\link{ToolExitCode}},
  \code{\link{ToolOutputReference}}
}

