% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{Execution}
\alias{Execution}
\title{Execution Object}
\usage{
Execution(completionTime = NULL, creationTime = NULL, executionId = NULL,
  outcome = NULL, state = NULL, testExecutionMatrixId = NULL)
}
\arguments{
\item{completionTime}{The time when the Execution status transitioned to COMPLETE}

\item{creationTime}{The time when the Execution was created}

\item{executionId}{A unique identifier within a History for this Execution}

\item{outcome}{Classify the result, for example into SUCCESS or FAILURE}

\item{state}{The initial state is IN_PROGRESS}

\item{testExecutionMatrixId}{TestExecution Matrix ID that the Test Service uses}
}
\value{
Execution object
}
\description{
Execution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Execution represents a collection of Steps. For instance, it could represent: - a mobile test executed across a range of device configurations - a jenkins job with a build step followed by a test stepThe maximum size of an execution message is 1 MiB.An Execution can be updated until its state is set to COMPLETE at which point it becomes immutable.
}
\seealso{
Other res: \code{\link{Any}}, \code{\link{Duration}},
  \code{\link{FailureDetail}}, \code{\link{FileReference}},
  \code{\link{History}}, \code{\link{Image}},
  \code{\link{InconclusiveDetail}},
  \code{\link{ListExecutionsResponse}},
  \code{\link{ListHistoriesResponse}},
  \code{\link{ListStepThumbnailsResponse}},
  \code{\link{ListStepsResponse}}, \code{\link{Outcome}},
  \code{\link{ProjectSettings}},
  \code{\link{PublishXunitXmlFilesRequest}},
  \code{\link{SkippedDetail}}, \code{\link{StackTrace}},
  \code{\link{Status}},
  \code{\link{StepDimensionValueEntry}},
  \code{\link{StepLabelsEntry}}, \code{\link{Step}},
  \code{\link{SuccessDetail}},
  \code{\link{TestCaseReference}},
  \code{\link{TestExecutionStep}}, \code{\link{TestIssue}},
  \code{\link{TestSuiteOverview}},
  \code{\link{TestTiming}}, \code{\link{Thumbnail}},
  \code{\link{Timestamp}}, \code{\link{ToolExecutionStep}},
  \code{\link{ToolExecution}}, \code{\link{ToolExitCode}},
  \code{\link{ToolOutputReference}}
}

