% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ListPerfSampleSeriesResponse}
\alias{ListPerfSampleSeriesResponse}
\title{ListPerfSampleSeriesResponse Object}
\usage{
ListPerfSampleSeriesResponse(perfSampleSeries = NULL)
}
\arguments{
\item{perfSampleSeries}{The resulting PerfSampleSeries sorted by id}
}
\value{
ListPerfSampleSeriesResponse object
}
\description{
ListPerfSampleSeriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
