% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ListStepsResponse}
\alias{ListStepsResponse}
\title{ListStepsResponse Object}
\usage{
ListStepsResponse(nextPageToken = NULL, steps = NULL)
}
\arguments{
\item{nextPageToken}{A continuation token to resume the query at the next item}

\item{steps}{Steps}
}
\value{
ListStepsResponse object
}
\description{
ListStepsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for StepService.List.
}
\seealso{
Other res: \code{\link{Any}}, \code{\link{Duration}},
  \code{\link{Execution}}, \code{\link{FailureDetail}},
  \code{\link{FileReference}}, \code{\link{History}},
  \code{\link{Image}}, \code{\link{InconclusiveDetail}},
  \code{\link{ListExecutionsResponse}},
  \code{\link{ListHistoriesResponse}},
  \code{\link{ListStepThumbnailsResponse}},
  \code{\link{Outcome}}, \code{\link{ProjectSettings}},
  \code{\link{PublishXunitXmlFilesRequest}},
  \code{\link{SkippedDetail}}, \code{\link{StackTrace}},
  \code{\link{Status}},
  \code{\link{StepDimensionValueEntry}},
  \code{\link{StepLabelsEntry}}, \code{\link{Step}},
  \code{\link{SuccessDetail}},
  \code{\link{TestCaseReference}},
  \code{\link{TestExecutionStep}}, \code{\link{TestIssue}},
  \code{\link{TestSuiteOverview}},
  \code{\link{TestTiming}}, \code{\link{Thumbnail}},
  \code{\link{Timestamp}}, \code{\link{ToolExecutionStep}},
  \code{\link{ToolExecution}}, \code{\link{ToolExitCode}},
  \code{\link{ToolOutputReference}}
}

