% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{MemoryInfo}
\alias{MemoryInfo}
\title{MemoryInfo Object}
\usage{
MemoryInfo(memoryCapInKibibyte = NULL, memoryTotalInKibibyte = NULL)
}
\arguments{
\item{memoryCapInKibibyte}{Maximum memory that can be allocated to the process in KiB}

\item{memoryTotalInKibibyte}{Total memory available on the device in KiB}
}
\value{
MemoryInfo object
}
\description{
MemoryInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
