% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{SkippedDetail}
\alias{SkippedDetail}
\title{SkippedDetail Object}
\usage{
SkippedDetail(incompatibleAppVersion = NULL,
  incompatibleArchitecture = NULL, incompatibleDevice = NULL)
}
\arguments{
\item{incompatibleAppVersion}{If the App doesn't support the specific API level}

\item{incompatibleArchitecture}{If the App doesn't run on the specific architecture, for example, x86}

\item{incompatibleDevice}{If the requested OS version doesn't run on the specific device model}
}
\value{
SkippedDetail object
}
\description{
SkippedDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other res: \code{\link{Any}}, \code{\link{Duration}},
  \code{\link{Execution}}, \code{\link{FailureDetail}},
  \code{\link{FileReference}}, \code{\link{History}},
  \code{\link{Image}}, \code{\link{InconclusiveDetail}},
  \code{\link{ListExecutionsResponse}},
  \code{\link{ListHistoriesResponse}},
  \code{\link{ListStepThumbnailsResponse}},
  \code{\link{ListStepsResponse}}, \code{\link{Outcome}},
  \code{\link{ProjectSettings}},
  \code{\link{PublishXunitXmlFilesRequest}},
  \code{\link{StackTrace}}, \code{\link{Status}},
  \code{\link{StepDimensionValueEntry}},
  \code{\link{StepLabelsEntry}}, \code{\link{Step}},
  \code{\link{SuccessDetail}},
  \code{\link{TestCaseReference}},
  \code{\link{TestExecutionStep}}, \code{\link{TestIssue}},
  \code{\link{TestSuiteOverview}},
  \code{\link{TestTiming}}, \code{\link{Thumbnail}},
  \code{\link{Timestamp}}, \code{\link{ToolExecutionStep}},
  \code{\link{ToolExecution}}, \code{\link{ToolExitCode}},
  \code{\link{ToolOutputReference}}
}

