% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{TestIssue}
\alias{TestIssue}
\title{TestIssue Object}
\usage{
TestIssue(errorMessage = NULL, stackTrace = NULL)
}
\arguments{
\item{errorMessage}{A brief human-readable message describing the abnormal event}

\item{stackTrace}{Optional}
}
\value{
TestIssue object
}
\description{
TestIssue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An abnormal event observed during the test execution.
}
\seealso{
Other res: \code{\link{Any}}, \code{\link{Duration}},
  \code{\link{Execution}}, \code{\link{FailureDetail}},
  \code{\link{FileReference}}, \code{\link{History}},
  \code{\link{Image}}, \code{\link{InconclusiveDetail}},
  \code{\link{ListExecutionsResponse}},
  \code{\link{ListHistoriesResponse}},
  \code{\link{ListStepThumbnailsResponse}},
  \code{\link{ListStepsResponse}}, \code{\link{Outcome}},
  \code{\link{ProjectSettings}},
  \code{\link{PublishXunitXmlFilesRequest}},
  \code{\link{SkippedDetail}}, \code{\link{StackTrace}},
  \code{\link{Status}},
  \code{\link{StepDimensionValueEntry}},
  \code{\link{StepLabelsEntry}}, \code{\link{Step}},
  \code{\link{SuccessDetail}},
  \code{\link{TestCaseReference}},
  \code{\link{TestExecutionStep}},
  \code{\link{TestSuiteOverview}},
  \code{\link{TestTiming}}, \code{\link{Thumbnail}},
  \code{\link{Timestamp}}, \code{\link{ToolExecutionStep}},
  \code{\link{ToolExecution}}, \code{\link{ToolExitCode}},
  \code{\link{ToolOutputReference}}
}

