% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{TestSuiteOverview}
\alias{TestSuiteOverview}
\title{TestSuiteOverview Object}
\usage{
TestSuiteOverview(errorCount = NULL, failureCount = NULL, name = NULL,
  skippedCount = NULL, totalCount = NULL, xmlSource = NULL)
}
\arguments{
\item{errorCount}{Number of test cases in error, typically set by the service by parsing the xml_source}

\item{failureCount}{Number of failed test cases, typically set by the service by parsing the xml_source}

\item{name}{The name of the test suite}

\item{skippedCount}{Number of test cases not run, typically set by the service by parsing the xml_source}

\item{totalCount}{Number of test cases, typically set by the service by parsing the xml_source}

\item{xmlSource}{If this test suite was parsed from XML, this is the URI where the original XML file is stored}
}
\value{
TestSuiteOverview object
}
\description{
TestSuiteOverview Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A summary of a test suite result either parsed from XML or uploaded directly by a user.Note: the API related comments are for StepService only. This message is also being used in ExecutionService in a read only mode for the corresponding step.
}
\seealso{
Other res: \code{\link{Any}}, \code{\link{Duration}},
  \code{\link{Execution}}, \code{\link{FailureDetail}},
  \code{\link{FileReference}}, \code{\link{History}},
  \code{\link{Image}}, \code{\link{InconclusiveDetail}},
  \code{\link{ListExecutionsResponse}},
  \code{\link{ListHistoriesResponse}},
  \code{\link{ListStepThumbnailsResponse}},
  \code{\link{ListStepsResponse}}, \code{\link{Outcome}},
  \code{\link{ProjectSettings}},
  \code{\link{PublishXunitXmlFilesRequest}},
  \code{\link{SkippedDetail}}, \code{\link{StackTrace}},
  \code{\link{Status}},
  \code{\link{StepDimensionValueEntry}},
  \code{\link{StepLabelsEntry}}, \code{\link{Step}},
  \code{\link{SuccessDetail}},
  \code{\link{TestCaseReference}},
  \code{\link{TestExecutionStep}}, \code{\link{TestIssue}},
  \code{\link{TestTiming}}, \code{\link{Thumbnail}},
  \code{\link{Timestamp}}, \code{\link{ToolExecutionStep}},
  \code{\link{ToolExecution}}, \code{\link{ToolExitCode}},
  \code{\link{ToolOutputReference}}
}

