% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{Thumbnail}
\alias{Thumbnail}
\title{Thumbnail Object}
\usage{
Thumbnail(contentType = NULL, data = NULL, heightPx = NULL,
  widthPx = NULL)
}
\arguments{
\item{contentType}{The thumbnail's content type, i}

\item{data}{The thumbnail file itself}

\item{heightPx}{The height of the thumbnail, in pixels}

\item{widthPx}{The width of the thumbnail, in pixels}
}
\value{
Thumbnail object
}
\description{
Thumbnail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single thumbnail, with its size and format.
}
\seealso{
Other res: \code{\link{Any}}, \code{\link{Duration}},
  \code{\link{Execution}}, \code{\link{FailureDetail}},
  \code{\link{FileReference}}, \code{\link{History}},
  \code{\link{Image}}, \code{\link{InconclusiveDetail}},
  \code{\link{ListExecutionsResponse}},
  \code{\link{ListHistoriesResponse}},
  \code{\link{ListStepThumbnailsResponse}},
  \code{\link{ListStepsResponse}}, \code{\link{Outcome}},
  \code{\link{ProjectSettings}},
  \code{\link{PublishXunitXmlFilesRequest}},
  \code{\link{SkippedDetail}}, \code{\link{StackTrace}},
  \code{\link{Status}},
  \code{\link{StepDimensionValueEntry}},
  \code{\link{StepLabelsEntry}}, \code{\link{Step}},
  \code{\link{SuccessDetail}},
  \code{\link{TestCaseReference}},
  \code{\link{TestExecutionStep}}, \code{\link{TestIssue}},
  \code{\link{TestSuiteOverview}},
  \code{\link{TestTiming}}, \code{\link{Timestamp}},
  \code{\link{ToolExecutionStep}},
  \code{\link{ToolExecution}}, \code{\link{ToolExitCode}},
  \code{\link{ToolOutputReference}}
}

