% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ToolExecutionStep}
\alias{ToolExecutionStep}
\title{ToolExecutionStep Object}
\usage{
ToolExecutionStep(toolExecution = NULL)
}
\arguments{
\item{toolExecution}{A Tool execution}
}
\value{
ToolExecutionStep object
}
\description{
ToolExecutionStep Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Generic tool step to be used for binaries we do not explicitly support. For example: running cp to copy artifacts from one location to another.
}
\seealso{
Other res: \code{\link{Any}}, \code{\link{Duration}},
  \code{\link{Execution}}, \code{\link{FailureDetail}},
  \code{\link{FileReference}}, \code{\link{History}},
  \code{\link{Image}}, \code{\link{InconclusiveDetail}},
  \code{\link{ListExecutionsResponse}},
  \code{\link{ListHistoriesResponse}},
  \code{\link{ListStepThumbnailsResponse}},
  \code{\link{ListStepsResponse}}, \code{\link{Outcome}},
  \code{\link{ProjectSettings}},
  \code{\link{PublishXunitXmlFilesRequest}},
  \code{\link{SkippedDetail}}, \code{\link{StackTrace}},
  \code{\link{Status}},
  \code{\link{StepDimensionValueEntry}},
  \code{\link{StepLabelsEntry}}, \code{\link{Step}},
  \code{\link{SuccessDetail}},
  \code{\link{TestCaseReference}},
  \code{\link{TestExecutionStep}}, \code{\link{TestIssue}},
  \code{\link{TestSuiteOverview}},
  \code{\link{TestTiming}}, \code{\link{Thumbnail}},
  \code{\link{Timestamp}}, \code{\link{ToolExecution}},
  \code{\link{ToolExitCode}},
  \code{\link{ToolOutputReference}}
}

