% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ToolOutputReference}
\alias{ToolOutputReference}
\title{ToolOutputReference Object}
\usage{
ToolOutputReference(creationTime = NULL, output = NULL, testCase = NULL)
}
\arguments{
\item{creationTime}{The creation time of the file}

\item{output}{A FileReference to an output file}

\item{testCase}{The test case to which this output file belongs}
}
\value{
ToolOutputReference object
}
\description{
ToolOutputReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference to a ToolExecution output file.
}
\seealso{
Other res: \code{\link{Any}}, \code{\link{Duration}},
  \code{\link{Execution}}, \code{\link{FailureDetail}},
  \code{\link{FileReference}}, \code{\link{History}},
  \code{\link{Image}}, \code{\link{InconclusiveDetail}},
  \code{\link{ListExecutionsResponse}},
  \code{\link{ListHistoriesResponse}},
  \code{\link{ListStepThumbnailsResponse}},
  \code{\link{ListStepsResponse}}, \code{\link{Outcome}},
  \code{\link{ProjectSettings}},
  \code{\link{PublishXunitXmlFilesRequest}},
  \code{\link{SkippedDetail}}, \code{\link{StackTrace}},
  \code{\link{Status}},
  \code{\link{StepDimensionValueEntry}},
  \code{\link{StepLabelsEntry}}, \code{\link{Step}},
  \code{\link{SuccessDetail}},
  \code{\link{TestCaseReference}},
  \code{\link{TestExecutionStep}}, \code{\link{TestIssue}},
  \code{\link{TestSuiteOverview}},
  \code{\link{TestTiming}}, \code{\link{Thumbnail}},
  \code{\link{Timestamp}}, \code{\link{ToolExecutionStep}},
  \code{\link{ToolExecution}}, \code{\link{ToolExitCode}}
}

