% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_functions.R
\docType{package}
\name{toolresults_googleAuthR}
\alias{toolresults_googleAuthR}
\alias{toolresults_googleAuthR-package}
\title{Cloud Tool Results API
Reads and publishes results from Cloud Test Lab.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:49:53
filename: /Users/mark/dev/R/autoGoogleAPI/googletoolresultsv1beta3.auto/R/toolresults_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}

